requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.6
django-cors-headers==4.2.0
django-filter==23.3
djangorestframework==3.14.0
drf-spectacular==0.26.5
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
Markdown==3.5
numpy==1.26.0
packaging==23.2
pandas==2.1.1
patsy==0.5.3
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.4
scipy==1.11.3
six==1.16.0
sqlparse==0.4.4
statsmodels==0.14.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
